Move image_bytes fixtures into top-level conftest
